Updated product price data type

+++ database/migrations/2021_09_07_062903_create_products_table.php
@@ -17,7 +17,7 @@ class CreateProductsTable extends Migration
             $table->id();
             $table->string('name')->unique();
             $table->text('description');
-            $table->float('price', 10, 2, true);
+            $table->decimal('price', 10, 2, true);
             $table->foreignId('product_category_id')->references('id')->on('product_categories');
             $table->timestamps();
         });
